@@ -1,10 +1,10 @@
 name: sign_in_with_kakao
 description: A new Flutter plugin.
-version: 0.1.1
+version: 1.0.0
 homepage: https://github.com/dev-juyoung/sign_in_with_kakao
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
