@@ -3,7 +3,7 @@ import '../types/types.dart';
 /// 카카오 로그인 예외
 class AuthFailureException implements Exception {
   final AuthFailureReason reason;
-  final String message;
+  final String? message;
 
   AuthFailureException(this.reason, this.message);
 
