@@ -1,5 +1,3 @@
-import '../extensions/extensions.dart';
-
 /// 사용자 정보 - 생일의 양력/음력 열거형
 enum BirthdayType {
   /// 양력
@@ -12,7 +10,7 @@ enum BirthdayType {
 extension BirthdayTypeExtension on BirthdayType {
   String get rawValue => this.toString().split('.').last;
 
-  static BirthdayType toBirthdayType(String rawValue) {
+  static BirthdayType? toBirthdayType(String? rawValue) {
     final value = rawValue != null ? rawValue.toLowerCase() : null;
 
     if (BirthdayType.solar.rawValue == value) {
