@@ -1,5 +1,3 @@
-import '../extensions/extensions.dart';
-
 /// 사용자 정보 - 성별 정보 열거형
 enum Gender {
   /// 남자
@@ -12,7 +10,7 @@ enum Gender {
 extension GenderExtension on Gender {
   String get rawValue => this.toString().split('.').last;
 
-  static Gender toGender(String rawValue) {
+  static Gender? toGender(String? rawValue) {
     final value = rawValue != null ? rawValue.toLowerCase() : null;
 
     if (Gender.male.rawValue == value) {
